KBMBF-452: #37 Show ratio of oer-content
Fixing bug of wrong node_id parameter in score
